requirements.txt
django
django-admin-tools
django-debug-toolbar
django-epiceditor
djangorestframework
feedparser
markdown
markdown2
mock
pillow
pygments
python-dateutil
